+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- [crawler] Enable usage of the GitHub user PRs crawler via the Monocle config.
+
 ### Changed
 
 ### Removed